requirements.txt
click==8.1.3
Flask==2.2.2
Flask-Login==0.6.2
Flask-MySQLdb==1.0.1
Flask-WTF==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
python-decouple==3.8
python-dotenv==1.0.0
Werkzeug==2.2.2
WTForms==3.0.1
